Rename ReutersEvaluator to ClassificationEvaluator
