+++ models/reg_lstm/__main__.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 import numpy as np
 import torch
 
-from common.evaluation import EvaluatorFactory
+from common.evaluate import EvaluatorFactory
 from common.train import TrainerFactory
 from datasets.aapd import AAPD
 from datasets.imdb import IMDB
